Complete Single tests
