+++ PasApiClient/ParsingException.cs
@@ -6,8 +6,11 @@ namespace PasApiClient
 {
     public class ParsingException : Exception
     {
-        public ParsingException(string message) : base(message)
+        public ParsingException(string message, int statusCode) : base(message)
         {
+            StatusCode = statusCode;
         }
+
+        public int StatusCode { get; }
     }
 }
